code_to_optimize/pie_test_set/tests/test_p02471.py
from code_to_optimize.pie_test_set.p02471 import problem_p02471


def test_problem_p02471_0():
    actual_output = problem_p02471("4 12")
    expected_output = "1 0"
    if isinstance(actual_output, type(expected_output)):
        assert actual_output == expected_output
    else:
        # Cast expected output to the type of actual output if they differ
        cast_expected_output = type(actual_output)(expected_output)
        assert actual_output == cast_expected_output


def test_problem_p02471_1():
    actual_output = problem_p02471("3 8")
    expected_output = "3 -1"
    if isinstance(actual_output, type(expected_output)):
        assert actual_output == expected_output
    else:
        # Cast expected output to the type of actual output if they differ
        cast_expected_output = type(actual_output)(expected_output)
        assert actual_output == cast_expected_output


def test_problem_p02471_2():
    actual_output = problem_p02471("4 12")
    expected_output = "1 0"
    if isinstance(actual_output, type(expected_output)):
        assert actual_output == expected_output
    else:
        # Cast expected output to the type of actual output if they differ
        cast_expected_output = type(actual_output)(expected_output)
        assert actual_output == cast_expected_output
